simplify `amount_split` by iterating over amount bits
Small optimization: rather than involving a string data type here (which
needs to also be reversed and sliced), we can just directly operate on
the amount integer variable by iterating over all the bits and add 2^i
to the return list if bit i is set.

Also, add type hints to the parameter / return value and assert that
the passed value must not be negative.